RF: require Python 2 for wheel build
It's too boring to work round the wheel dynamic library filenames for
different Python versions - pin to Python 2 for now.

+++ wheel_makers/make_wheels.sh
@@ -1,8 +1,15 @@
 # Make wheels and copy into main package
 # This is to build the wheels we use for testing
-# Need cython and wheel installed to run this script
-# Use wheel==0.23 for compatibility with tests.
+# Need Cython and wheel installed to run this script
 # Run on earliest supported version of OSX (currently 10.6)
+
+# Always use Python 2 (for consistent dynamic lib extensions)
+py_major=$(python -c 'import sys; print(sys.version_info[0])')
+if [ "$py_major" != 2 ]; then
+    echo Need Python 2 for wheel building
+    exit 1
+fi
+
 rm */dist/fakepkg*.whl
 rm */libs/*.dylib
